Adding Layouts
Add a 2x2 grid of buttons to the main screen, where each button displays
a 'toast' when tapped.

When the top-left cell is tapped, take the user to a new screen that
displays a vertically scrolling list of text information. Each list row
should have 2+ pieces of information (e.g. book title & author) in
separate views. The list uses a RecyclerView to display more items than
can fit on the screen at once.